Add Winston 3.x support
