+++ README.md
@@ -1,6 +1,6 @@
 # @cardash/winston-datadog-tcp [![CircleCI](https://circleci.com/gh/cardash/winston-datadog-tcp/tree/master.svg?style=svg)](https://circleci.com/gh/cardash/winston-datadog-tcp/tree/master) [![Coverage Status](https://coveralls.io/repos/github/cardash/winston-datadog-tcp/badge.svg?branch=master)](https://coveralls.io/github/cardash/winston-datadog-tcp?branch=master)  [![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)](https://opensource.org/licenses/MIT)
 
-A Winston-compatible (2.x) Datadog logger via TCP+SSL. No local file taking up space, no log rotation management.
+A Winston-compatible (3.x and 2.x) Datadog logger via TCP+SSL. No local file taking up space, no log rotation management.
 
 ## Install
 
@@ -8,11 +8,41 @@ A Winston-compatible (2.x) Datadog logger via TCP+SSL. No local file taking up s
 
 ## Usage
 
+This library supports both Winston 3.x and 2.x, with 3.x being the default.
+
+### Winston 3.x
+
 Just provide the library your Datadog API Key and attach it as a Winston transport. Optionally add datadog tags to include in object notation. The `app` tag is also used to flag the service to Datadog for APM integration if present.
 
 ```js
 const config = require('@cardash/config')
-const winstonDatadogTcp = require('@cardash/winston-datadog-tcp')
+const WinstonDatadogTcp = require('@cardash/winston-datadog-tcp')
+const winston = require('winston')
+
+const logger = new winston.createLogger({
+  level: config.logLevel,
+  transports: [
+    new winston.transports.Console(),
+    new WinstonDatadogTcp({
+      apiKey: config.datadogKey,
+      tags: {
+        app: 'my-server',
+        env: config.NODE_ENV,
+      },
+    }),
+  ],
+})
+
+...
+```
+
+### Winston 2.x
+
+Just provide the library your Datadog API Key and attach it as a Winston transport. Optionally add datadog tags to include in object notation. The `app` tag is also used to flag the service to Datadog for APM integration if present.
+
+```js
+const config = require('@cardash/config')
+const winstonDatadogTcp = require('@cardash/winston-datadog-tcp/2.x')
 const winston = require('winston')
 
 const logger = new winston.Logger({
