+++ index.js
@@ -1,55 +1,21 @@
-const tls = require('tls')
-const hostname = require('os').hostname()
+const BaseTransport = require('winston-transport')
+const winstonDatadogTcp = require('./2.x')
 
-const createConnection = () => {
-  const conn = tls.connect({
-    host: 'intake.logs.datadoghq.com',
-    port: 10516,
-  })
-  conn.setEncoding('utf8')
-  conn.setKeepAlive(true)
-  conn.unref()
-  conn.on('error', () => { conn.destroyed = true })
-
-  return conn
-}
-
-const winstonDatadogTcp = (apiKey, tags) => {
-  const dd = {
-    createConnection,
-    conn: createConnection(),
-    queuedMessage: undefined,
+class WinstonDatadogTcp extends BaseTransport {
+  constructor(opts) {
+    super(opts)
+    const { apiKey, tags, } = opts
+    this.coreLogger = winstonDatadogTcp(apiKey, tags)
   }
-  const ddtags = Object.keys(tags).map(t => `${t}:${tags[t]}`).join(',')
-  return {
-    dd,
-    log: (loglevel, text, data, callback) => {
-      const record = {
-        status: loglevel,
-        message: `${text} ${JSON.stringify(data)}`,
-        data,
-        ddtags,
-        ddsource: '@cardash/winston-datadog-tcp',
-        hostname,
-        service: tags.app || undefined,
-      }
-      if (dd.conn.destroyed) {
-        dd.conn = dd.createConnection()
-        if (dd.queuedMessage) return dd.conn.write(dd.queuedMessage, () => {
-          dd.queuedMessage = `${apiKey} ${JSON.stringify(record)}\n`
-          dd.conn.write(dd.queuedMessage, () => {
-            dd.queuedMessage = undefined
-            callback()
-          })
-        })
-      }
-      dd.queuedMessage = `${apiKey} ${JSON.stringify(record)}\n`
-      dd.conn.write(dd.queuedMessage, () => {
-        dd.queuedMessage = undefined
-        callback()
-      })
-    }
+
+  log(info, callback) {
+    const {
+      level,
+      message,
+      ...data
+    } = info
+    this.coreLogger.log(level, message, data, callback)
   }
 }
 
-module.exports = winstonDatadogTcp
+module.exports = WinstonDatadogTcp
