+++ package.json
@@ -30,5 +30,8 @@
     "eslint": "^6.0.1",
     "nodeunit": "^0.11.3",
     "nyc": "^14.1.1"
+  },
+  "dependencies": {
+    "winston-transport": "^4.3.0"
   }
 }
